home-estimator/requirements.txt
torch>=2.0
torchvision>=0.15
scikit-learn>=1.3
streamlit>=1.28
Pillow>=10.0
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
spacy>=3.6
# After install: python -m spacy download en_core_web_sm
wordcloud>=1.9
icrawler>=0.6
joblib>=1.3
openai>=1.0
openai-whisper>=20230918
streamlit-audiorec>=0.1
python-dotenv>=1.0
praw>=7.7
